+++ src/main/java/com/team/buddyya/feed/controller/CommentController.java
@@ -1,15 +1,17 @@
 package com.team.buddyya.feed.controller;
 
 import com.team.buddyya.auth.domain.CustomUserDetails;
-import com.team.buddyya.feed.dto.request.CommentCreateRequest;
-import com.team.buddyya.feed.dto.request.CommentUpdateRequest;
-import com.team.buddyya.feed.dto.response.CommentCreateResponse;
-import com.team.buddyya.feed.dto.response.CommentUpdateResponse;
+import com.team.buddyya.feed.dto.request.comment.CommentCreateRequest;
+import com.team.buddyya.feed.dto.request.comment.CommentUpdateRequest;
+import com.team.buddyya.feed.dto.response.comment.CommentCreateResponse;
+import com.team.buddyya.feed.dto.response.comment.CommentListResponse;
+import com.team.buddyya.feed.dto.response.comment.CommentUpdateResponse;
 import com.team.buddyya.feed.service.CommentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -23,6 +25,13 @@ import org.springframework.web.bind.annotation.RestController;
 public class CommentController {
 
     private final CommentService commentService;
+
+    @GetMapping
+    public ResponseEntity<CommentListResponse> getComments(@AuthenticationPrincipal CustomUserDetails customUserDetails,
+                                                           @PathVariable Long feedId) {
+        CommentListResponse response = commentService.getComments(customUserDetails.getStudentInfo(), feedId);
+        return ResponseEntity.ok(response);
+    }
 
     @PostMapping
     public ResponseEntity<CommentCreateResponse> createComment(@AuthenticationPrincipal CustomUserDetails userDetails,
