+++ src/main/java/com/team/buddyya/feed/dto/request/comment/CommentCreateRequest.java
@@ -1,4 +1,4 @@
-package com.team.buddyya.feed.dto.request;
+package com.team.buddyya.feed.dto.request.comment;
 
 public record CommentCreateRequest(String content) {
 }
