+++ src/main/java/com/team/buddyya/feed/dto/request/comment/CommentUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.team.buddyya.feed.dto.request;
+package com.team.buddyya.feed.dto.request.comment;
 
 public record CommentUpdateRequest(String content) {
 }
