+++ src/main/java/com/team/buddyya/feed/dto/request/feed/FeedListRequest.java
@@ -1,4 +1,4 @@
-package com.team.buddyya.feed.dto.request;
+package com.team.buddyya.feed.dto.request.feed;
 
 public record FeedListRequest(
         int page,
