+++ src/main/java/com/team/buddyya/feed/dto/response/comment/CommentUpdateResponse.java
@@ -1,4 +1,4 @@
-package com.team.buddyya.feed.dto.response;
+package com.team.buddyya.feed.dto.response.comment;
 
 import com.team.buddyya.feed.domain.Comment;
 import java.time.LocalDateTime;
@@ -10,7 +10,7 @@ public record CommentUpdateResponse(
         String country,
         LocalDateTime createdDate
 ) {
-    
+
     public static CommentUpdateResponse from(Comment comment) {
         return new CommentUpdateResponse(
                 comment.getId(),
