+++ src/main/java/com/team/buddyya/feed/dto/response/feed/FeedCreateResponse.java
@@ -1,7 +1,7 @@
-package com.team.buddyya.feed.dto.response;
+package com.team.buddyya.feed.dto.response.feed;
 
 public record FeedCreateResponse(Long id) {
-    
+
     public static FeedCreateResponse from(Long id) {
         return new FeedCreateResponse(id);
     }
