+++ src/main/java/com/team/buddyya/feed/service/CommentService.java
@@ -3,12 +3,12 @@ package com.team.buddyya.feed.service;
 import com.team.buddyya.auth.domain.StudentInfo;
 import com.team.buddyya.feed.domain.Comment;
 import com.team.buddyya.feed.domain.Feed;
-import com.team.buddyya.feed.dto.request.CommentCreateRequest;
-import com.team.buddyya.feed.dto.request.CommentUpdateRequest;
-import com.team.buddyya.feed.dto.response.CommentCreateResponse;
-import com.team.buddyya.feed.dto.response.CommentListResponse;
-import com.team.buddyya.feed.dto.response.CommentResponse;
-import com.team.buddyya.feed.dto.response.CommentUpdateResponse;
+import com.team.buddyya.feed.dto.request.comment.CommentCreateRequest;
+import com.team.buddyya.feed.dto.request.comment.CommentUpdateRequest;
+import com.team.buddyya.feed.dto.response.comment.CommentCreateResponse;
+import com.team.buddyya.feed.dto.response.comment.CommentListResponse;
+import com.team.buddyya.feed.dto.response.comment.CommentResponse;
+import com.team.buddyya.feed.dto.response.comment.CommentUpdateResponse;
 import com.team.buddyya.feed.exception.FeedException;
 import com.team.buddyya.feed.exception.FeedExceptionType;
 import com.team.buddyya.feed.respository.CommentRepository;
