+++ src/main/java/com/team/buddyya/feed/service/FeedService.java
@@ -4,12 +4,12 @@ import com.team.buddyya.auth.domain.StudentInfo;
 import com.team.buddyya.feed.domain.Category;
 import com.team.buddyya.feed.domain.Feed;
 import com.team.buddyya.feed.domain.FeedUserAction;
-import com.team.buddyya.feed.dto.request.FeedCreateRequest;
-import com.team.buddyya.feed.dto.request.FeedListRequest;
-import com.team.buddyya.feed.dto.request.FeedUpdateRequest;
-import com.team.buddyya.feed.dto.response.FeedListItemResponse;
-import com.team.buddyya.feed.dto.response.FeedListResponse;
-import com.team.buddyya.feed.dto.response.FeedResponse;
+import com.team.buddyya.feed.dto.request.feed.FeedCreateRequest;
+import com.team.buddyya.feed.dto.request.feed.FeedListRequest;
+import com.team.buddyya.feed.dto.request.feed.FeedUpdateRequest;
+import com.team.buddyya.feed.dto.response.feed.FeedListItemResponse;
+import com.team.buddyya.feed.dto.response.feed.FeedListResponse;
+import com.team.buddyya.feed.dto.response.feed.FeedResponse;
 import com.team.buddyya.feed.exception.FeedException;
 import com.team.buddyya.feed.exception.FeedExceptionType;
 import com.team.buddyya.feed.respository.FeedRepository;
